WIP: use record_details endpoint to render pin modal
