+++ src/constants.ts
@@ -12,7 +12,7 @@ export const all_columns = ["source_id", "state", "city", "country", "age", "ser
 "sampling_start_date", "sampling_end_date", "url", "first_author", "source_type", "study_type", "test_type", "specimen_type",
 "isotypes_reported", "approving_regulator", "population_group", "lead_organization"];
 
-export const explore_modal_columns = ["source_id", "state", "city", "country", "age", "serum_pos_prevalence", "denominator_value", "population_group", 
+export const pins_modal_columns = ["source_id", "state", "city", "country", "age", "serum_pos_prevalence", "denominator_value", "population_group", 
 "overall_risk_of_bias", "source_name", "estimate_grade", "pin_latitude", "pin_longitude", "pin_region_type", 
 "sampling_start_date", "sampling_end_date", "url"];
